# docs/swagger.yaml
swagger: '2.0'
info:
  title: Somosport Core API
  description: Basic description of services provided by Somosport Core. Online editor & validator available at  http://editor.swagger.io/
  version: "1.0"
# the domain of the service
# host: ss-core-dev.herokuapp.com
#host: localhost:3000
# array of all schemes that your API supports
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /api/v1.0
produces:
  - application/json

##################################################
#
#  _____                 _
# /  ___|               (_)
# \ `--.  ___ _ ____   ___  ___ ___  ___
#  `--. \/ _ \ '__\ \ / / |/ __/ _ \/ __|
# /\__/ /  __/ |   \ V /| | (_|  __/\__ \
# \____/ \___|_|    \_/ |_|\___\___||___/
#
#
###################################################
paths:
...
  /category/{category_id}/team/{team_id}/player:
    get:
      summary: Gets all the players that participate on a team in a category. Also known as summoned players
      description: Gets all the players that participate on a team in a category. Also knownas summoned players
      operationId: get_category_team_player_
      produces:
      - application/json
      parameters:
        - name: category_id
          in: path
          required: true
          type: integer
        - name: team_id
          in: path
          required: true
          type: integer
      tags:
        - Player
        - Category
        - Team
      responses:
        200:
          description: pending
  /category/{category_id}/team/{team_id}/player/{player_id}:
    get:
      summary: REVISAR Gets the players that participate on a team in a category
      description: duplicado? Get players by category_id and team_id
      operationId: get_category_team_player
      produces:
      - application/json
      parameters:
        - name: category_id
          in: path
          required: true
          type: integer
        - name: team_id
          in: path
          required: true
          type: integer
        - name: player_id
          in: path
          required: true
          type: integer
          format: integer
      tags:
        - Player
        - Category
        - Team
      responses:
        200:
          description: pending
    post:
      summary: Create info of the summoned player of one team on one category
      description: Create info of the summoned player of one team on one category
      operationId: post_category_team_player
      tags:
        - Player
        - Category
        - Team
      produces:
      - application/json
      parameters:
      - name: category_id
        in: path
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        required: true
        type: integer
        format: integer
      - name: player_id
        in: path
        required: true
        type: integer
        format: integer
      - name: Body
        in: body
        required: false
        x-is-map: false
        schema:
          $ref: '#/definitions/CreateCategoryTeamPlayer'
      responses:
        200:
          description: Create info of the summoned player of one team on one category successful
      security: []
    put:
      summary: Update info of the summoned player of one team on one category
      description: Update info of the summoned player of one team on one category
      tags:
        - Player
        - Category
        - Team
      produces:
      - application/json
      parameters:
      - name: category_id
        in: path
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        required: true
        type: integer
        format: integer
      - name: player_id
        in: path
        required: true
        type: integer
        format: integer
      - name: Body
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/UpdateCategyTeamPlayer'
      responses:
        200:
          description: Update info of the summoned player of one team on one category successful
      security: []
...
  /match/{match_id}/player:
    get:
      summary: Returns the player list for this match, grouped by home_team & visitor_team
      description: Returns the player list for this match, grouped by home_team & visitor_team
      tags:
        - Match
        - Player
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: list of players available for this match
  /match/{match_id}/team/{team_id}/player:
    get:
      summary: Gets the player list for this team & match
      description: Gets the player list for this team & match
      tags:
        - Match
        - Player
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: List of players assigned to this match & team
    put:
      summary: Toggles the participation of the given player in this match
      description: Toggles the participation of the given player in this match; necessary for updating the summoned player list for each match
      tags:
        - Match
        - Player
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      - name: player_data
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/MatchPlayerTeamRequest'
      responses:
        200:
          description: Correct update of match player list
...
  /match/:match_id/team/:team_id/player:
    post:
      summary: registers the starter eleven players for a match for each team
      description: registers the starter eleven players for a match for each team
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      - name: events
        in: body
        required: true
        schema:
          $ref: '#/definitions/initialPlayers'
      responses:
        200:
          description: (documentation pending)
...
  /team/{team_id}/player:
    get:
      summary: Gets the team's roster of players
      description: Gets the team's roster of players.
      parameters:
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
      tags:
        - Team
        - Player
      responses:
        200:
          description: An array of players
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Creates a player and inserts it in the team roster
      description: Receives an object or an array of objects, as described in the model example
      parameters:
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
        - name: player & team data
          in: body
          required: true
          schema:
            $ref: '#/definitions/PlayerTeamData'
      tags:
        - Team
        - Player
      responses:
        200:
          description: result
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  # /team/{team_id}/player/{player_id}:
  #   get:
  #     summary: Get the team's player
  #     description: Get the team's player
  #     parameters:
  #       - name: team_id
  #         in: path
  #         description: Team ID
  #         required: true
  #         type: integer
  #         format: integer
  #       - name: player_id
  #         in: path
  #         description: PLAYER ID
  #         required: true
  #         type: integer
  #         format: integer
  #     tags:
  #       - Team
  #       - Player
  #     responses:
  #       200:
  #         description: An team players
  #       default:
  #         description: Unexpected error
  #         schema:
  #           $ref: '#/definitions/Error'
  #   put:
  #     summary: Updates a player and inserts it in the team roster
  #     description: Updates a player and inserts it in the team roster
  #     parameters:
  #       - name: team_id
  #         in: path
  #         description: Team ID
  #         required: true
  #         type: integer
  #         format: integer
  #       - name: player and team data
  #         in: body
  #         required: true
  #         x-is-map: false
  #         schema:
  #           $ref: '#/definitions/PlayerTeamRequest'
  #     tags:
  #       - Team
  #       - Player
  #     responses:
  #       200:
  #         description: result
  #       default:
  #         description: Unexpected error
  #         schema:
  #           $ref: '#/definitions/Error'
  #   delete:
  #     summary: Sets a player of the roster as inactive. It does *not* deactivates the player
  #     description: Sets a player of the roster as inactive. It does *not* deactivates the player
  #     parameters:
  #       - name: team_id
  #         in: path
  #         description: Team ID
  #         required: true
  #         type: integer
  #         format: integer
  #       - name: player_id
  #         in: path
  #         required: true
  #         type: integer
  #     tags:
  #       - Team
  #       - Player
  #     responses:
  #       200:
  #         description: result
  #       default:
  #         description: Unexpected error
  #         schema:
  #           $ref: '#/definitions/Error'
...
  /player/{player_id}:
    get:
      summary: Gets player's info
      description: Gets player's info
      parameters:
        - name: player_id
          in: path
          required: true
          type: integer
      tags:
        - Player
      responses:
        200:
          description: Information about a single player
...
  /player/{player_id}/category/{category_id}/events:
    get:
      summary: Gets all player's events in a specific category
      description: Gets all player's events in a specific category
      parameters:
        - name: category_id
          in: path
          required: true
          type: integer
        - name: player_id
          in: path
          required: true
          type: integer
      tags:
        - Player
      responses:
        200:
          description: List of events
...
  /person:
    get:
      summary: Gets the info of all persons
      description: Gets the info of all persons
      tags:
        - Person
      responses:
        200:
          description: List of persons
    post:
      summary: Creates a person
      description: Creates a person
      tags:
        - Person
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/PersonRequest'
      responses:
        200:
          description: person creation OK
      security: []
  /person/{person_id}:
    get:
      summary: Returns the detail of one person
      description: Returns the detail of one person
      parameters:
        - name: person_id
          in: path
          description: Person ID
          required: true
          type: integer
          format: integer
      tags:
        - Person
      responses:
        200:
          description: An array of persons
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Updates a person
      description: Updates a person
      parameters:
        - name: Authorization-Token
          in: header
          type: string
          required: true
        - name: person_id
          in: path
          description: Person ID
          required: true
          type: integer
          format: integer
        - name: person data
          in: body
          required: true
          x-is-map: false
          schema:
            $ref: '#/definitions/PersonRequest'
      tags:
        - Person
      responses:
        200:
          description: Resource updated
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'


#######################################################################
#      _       __ _       _ _   _
#     | |     / _(_)     (_) | (_)
#   __| | ___| |_ _ _ __  _| |_ _  ___  _ __  ___
#  / _` |/ _ \  _| | '_ \| | __| |/ _ \| '_ \/ __|
# | (_| |  __/ | | | | | | | |_| | (_) | | | \__ \
#  \__,_|\___|_| |_|_| |_|_|\__|_|\___/|_| |_|___/
#
#######################################################################
definitions:
...
  CreateGroupRequest:
    title: Match Creation Request
    type: object
    properties:
      name:
        type: string
      phase_id:
        type: integer
      participant_team:
        type: integer
      classified_team:
        type: integer
    required:
    - name
    - phase_id
  MatchPlayerTeamRequest:
    title: Match Creation Request
    type: object
    properties:
      position:
        type: string
      number:
        type: integer
      player_id:
        type: integer
    required:
    - player_id
  CreateTeamRequest:
    title: Team Creation/Update Request
    type: object
    properties:
      name:
        type: string
      logo_url:
        type: string
      short_name:
        type: string
      description:
        type: string
      category_type_id:
        type: integer
        description: If not provided, a new 'club' organization will be created under the same name as the team
      gender_id:
        type: integer
        description: Gender associated to this team
      organization_id:
        type: integer
        description: If not provided, a new 'club' organization will be created under the same name as the team
      subdiscipline_id:
        type: integer
        description: Sport associated with this team
    required:
      - name
      - category_type_id
      - gender_id
      - subdiscipline_id
...
  PlayerTeamRequest:
    title: PlayerTeam Request
    type: object
    properties:
      player:
        $ref: '#/definitions/CreateTeamRequest'
      team:
        $ref: '#/definitions/CreateGroupRequest'
  PlayerTeamData:
    type: array
    items:
      type: object
      properties:
        player:
          type: object
          description: player data
          required: true
          properties:
            first_name:
              type: string
              description: First name of the player
              required: true
            last_name:
              type: string
              description: Last name of the player
              required: true
            birthday:
              type: string
              description: Player's date of birth. ex. 2001-02-12T00:00:00-04:00
              required: true
            nickname:
              type: string
              description: Player's nickname or familiar name
              required: false
            email:
              type: string
              description: Player's email
              required: false
            document_number:
              type: string
              description: Identification number of the player, if available
              required: false
        team_player:
          type: object
          required: true
          properties:
            number:
              type: integer
              description: Player's number on team
              required: true
            position_id:
              type: integer
              description: Player's position on the team
              required: true
            team_id:
              type: integer
              description: Team identifier
              required: true
...
  UpdateCategyTeamPlayer:
    title: UpdateCategyTeamPlayer
    type: object
    properties:
      id:
        type: integer
        format: int32
      active:
        type: boolean
      number:
        type: integer
        format: int32
      position:
        type: string
    required:
    - id
  CreateCategoryTeamPlayer:
    title: CreateCategoryTeamPlayer
    type: object
    properties:
      category_id:
        type: string
      team_id:
        type: string
      player_id:
        type: string
      active:
        type: boolean
    required:
    - category_id
    - team_id
    - player_id
    - active
    - number
    - position
...
  initialPlayers:
      title: initialPlayers
      type: array
      items:
        type: object
        properties:
          number:
            type: integer
          player_id:
            type: integer
          match_id:
            type: integer
          team_id:
            type: integer
  Error:
    title: error
    type: object
    properties:
      description:
        type: string
...
  PersonRequest:
    title: Person Creation/Update Request
    type: object
    properties:
      name:
        type: string
      last_name:
        type: string
      nickname:
        type: string
      email:
        type: string
      birthday:
        type: string
        description: Ex. 2001-02-12T00:00:00-04:00
      height:
        type: string
      weight:
        type: string
      img_url:
        type: string
      document_number:
        type: string
      document_img_url:
        type: string
      nationality:
        type: string
      claimed:
        type: string
      active:
        type: boolean
      gender_id:
        type: integer
      status_type_id:
        type: integer
      meta:
        type: string
...